Authorization Complete

+++ node-server/routes/posts.js
@@ -36,7 +36,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -67,11 +68,16 @@ router.put(
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: "Update successful!"});
+  // console.log(post);
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: "Update successful!"});
+    } else {
+      res.status(401).json({message: "Not Authorized!"});
+    }
   });
  }
 );
@@ -112,9 +118,13 @@ router.get("/:id", (req, res, next) => { // called from "client" posts.service
 
 router.delete("/:id", confirmAuth, // executed in compare-auth.js
   (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    // console.log(result);
+    if (result.n > 0) {
+      res.status(200).json({message: "Deleted successfully!"});
+    } else {
+      res.status(401).json({message: "Not Authorized!"});
+    }
   });
 });
